order-api/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = []
dependencies = [
    "fastapi>=0.101.1",
    "uvicorn>=0.23.2",
    "boto3>=1.28.32",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

